+++ speeduino/src/decoders/420a.cpp
@@ -166,7 +166,7 @@ int getCrankAngle_420a(void)
   //This is the current angle ATDC the engine is at. This is the last known
   //position based on what tooth was last 'seen'. It is only accurate to the
   //resolution of the trigger wheel (Eg 36-1 is 10 degrees)
-  unsigned long tempToothLastToothTime;
+  uint32_t tempToothLastToothTime;
   int tempToothCurrentCount;
   //Grab some variables that are used in the trigger code and assign them to temp variables.
 
@@ -174,7 +174,7 @@ int getCrankAngle_420a(void)
 
   tempToothCurrentCount = toothCurrentCount;
   tempToothLastToothTime = toothLastToothTime;
-  unsigned long const lastCrankAngleCalc = micros();
+  uint32_t const lastCrankAngleCalc = micros();
 
   interrupts();
